Statistics handler as an argument of JFAdapter
